Add Validation for SuperGroupAdminController

+++ backend/src/main/java/it/chalmers/gamma/controller/admin/SuperGroupAdminController.java
@@ -2,24 +2,22 @@ package it.chalmers.gamma.controller.admin;
 
 import it.chalmers.gamma.db.entity.FKITSuperGroup;
 import it.chalmers.gamma.requests.CreateSuperGroupRequest;
-import it.chalmers.gamma.response.FKITSuperGroupCreatedResponse;
-import it.chalmers.gamma.response.GetGroupsResponse;
-import it.chalmers.gamma.response.GetSuperGroupResponse;
-import it.chalmers.gamma.response.GroupAlreadyExistsResponse;
-import it.chalmers.gamma.response.GroupDeletedResponse;
-import it.chalmers.gamma.response.GroupDoesNotExistResponse;
-import it.chalmers.gamma.response.GroupEditedResponse;
+import it.chalmers.gamma.response.*;
 import it.chalmers.gamma.service.FKITSuperGroupService;
 
 import java.util.List;
 import java.util.UUID;
 
+import it.chalmers.gamma.util.InputValidationUtils;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+
+import javax.validation.Valid;
 
 
 @RestController
@@ -34,7 +32,11 @@ public class SuperGroupAdminController {
     }
 
     @RequestMapping(method = RequestMethod.POST)
-    public ResponseEntity<FKITSuperGroup> createSuperGroup(@RequestBody CreateSuperGroupRequest request) {
+    public ResponseEntity<FKITSuperGroup> createSuperGroup(@Valid @RequestBody CreateSuperGroupRequest request,
+                                                           BindingResult result) {
+        if(result.hasErrors()){
+            throw new InputValidationFailedResponse(InputValidationUtils.getErrorMessages(result.getAllErrors()));
+        }
         if (this.fkitSuperGroupService.groupExists(request.getName())) {
             throw new GroupAlreadyExistsResponse();
         }
@@ -48,12 +50,18 @@ public class SuperGroupAdminController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     public ResponseEntity<FKITSuperGroup> getSuperGroup(@PathVariable("id") String id) {
+        if(!this.fkitSuperGroupService.groupExists(UUID.fromString(id))){
+            throw new GroupDoesNotExistResponse();
+        }
         return new GetSuperGroupResponse(this.fkitSuperGroupService.getGroup(UUID.fromString(id)));
     }
 
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<String> removeSuperGroup(@PathVariable("id") String id) {
+        if(!this.fkitSuperGroupService.groupExists(UUID.fromString(id))){
+            throw new GroupDoesNotExistResponse();
+        }
         this.fkitSuperGroupService.removeGroup(UUID.fromString(id));
         return new GroupDeletedResponse();
     }
